% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding-list.R
\name{as_coding_list}
\alias{as_coding_list}
\title{Evaluate a collection of codings from a character vector}
\usage{
as_coding_list(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A list of codings
}
\description{
Evaluate a collection of codings from a character vector
}
\examples{
char_vec <- c("coding(code('Yes', 1), code('No', 0))", "")
as_coding_list(char_vec)
}
