% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoding.R
\name{assign_coding}
\alias{assign_coding}
\title{Adds a coding as an attribute to a vector}
\usage{
assign_coding(vec, .coding)
}
\arguments{
\item{vec}{A vector}

\item{.coding}{A `coding` object}
}
\value{
The vector with its "rcoder.coding" attribute set to `.coding`
}
\description{
Stores a coding at the "rcoder.coding" attribute of a vector
}
