% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{coding_to_haven_labels}
\alias{coding_to_haven_labels}
\title{Convert coding to `haven`-compatible labels}
\usage{
coding_to_haven_labels(coding)
}
\arguments{
\item{coding}{A coding object}
}
\value{
A named vector representation of the coding
}
\description{
Converts a `coding` object into a named vector to be used in the `labels`
parameter for `haven::labelled()`.
}
