% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{eval_coding}
\alias{eval_coding}
\title{Evaluates a coding expression in a safe environment}
\usage{
eval_coding(expr)
}
\arguments{
\item{expr}{An expression}
}
\value{
An evaluated `coding` object
}
\description{
To prevent requiring attaching the `rcoder` package, this function takes in
an unevaluated expression -- assumed to be a `coding()` call -- and evaluates
the expression with _only_ `coding` and `code` provided to guard against
rogue code.
}
