% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{is_empty_coding}
\alias{is_empty_coding}
\title{Is an object the empty coding?}
\usage{
is_empty_coding(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE/FALSE if the object is identical to `empty_coding()`
}
\description{
Is an object the empty coding?
}
