% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\name{missing_codes}
\alias{missing_codes}
\title{Get missing codes from a coding}
\usage{
missing_codes(coding)
}
\arguments{
\item{coding}{a coding}
}
\value{
A coding that contains all missing codes.
  If no codes are found, returns `empty_coding()`
}
\description{
Takes a coding a returns a new coding
with all codes that represent a missing value.
}
\examples{
missing_codes(coding(code("Yes", 1), code("No", 0), code("Missing", NA)))
missing_codes(coding(code("Yes", 1), code("No", 0)))
}
