% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_codings}
\alias{link_codings}
\title{Link a coding from others for recoding}
\usage{
link_codings(to, from, to_suffix = "to", drop_unused = TRUE)
}
\arguments{
\item{to}{A coding to be linked to}

\item{from}{A coding or list of codings to be linked from}

\item{to_suffix}{A suffix signifying which columns in the output `data.frame`
came from `to`}

\item{drop_unused}{Logical flag to drop any codes in `to` that have no
counterparts in `from`}
}
\value{
A `linked_coding_df` with all necessary information for a recoding
  query
}
\description{
Coding objects can be linked together to create mappings from one or more
codings to another. This creates a `data.frame` that outlines how the codings
are linked, to be used in `make_recode_query()`.
}
